Refactor proof of work related functions out of main
