+++ src/miner.cpp
@@ -9,6 +9,7 @@
 #include "hash.h"
 #include "main.h"
 #include "net.h"
+#include "pow.h"
 #ifdef ENABLE_WALLET
 #include "wallet.h"
 #endif
