+++ src/rpcmining.cpp
@@ -9,6 +9,7 @@
 #include "net.h"
 #include "main.h"
 #include "miner.h"
+#include "pow.h"
 #ifdef ENABLE_WALLET
 #include "db.h"
 #include "wallet.h"
