+++ src/test/DoS_tests.cpp
@@ -11,6 +11,7 @@
 #include "keystore.h"
 #include "main.h"
 #include "net.h"
+#include "pow.h"
 #include "script.h"
 #include "serialize.h"
